// src/app/courses/interfaces/course-turn.interface.ts
// Generated by https://quicktype.io

export interface CourseTurn {
  id: number;
  course_id: number;
  turn_id: number;
  start_hour: string;
  end_hour: string;
  days: string;
  created_at: null;
  updated_at: null;
  turn: Turn;
}

export interface Turn {
  id: number;
  name: string;
  created_at: string;
  updated_at: string;
}